+++ noodles-fastq/src/record.rs
@@ -4,8 +4,6 @@ mod definition;
 
 pub use self::definition::Definition;
 
-use std::fmt;
-
 /// A FASTQ record.
 #[derive(Clone, Default, Debug, Eq, PartialEq)]
 pub struct Record {
@@ -167,54 +165,9 @@ impl Record {
     }
 }
 
-impl fmt::Display for Record {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str("@")?;
-
-        for &b in self.name() {
-            write!(f, "{}", b as char)?;
-        }
-
-        if !self.description().is_empty() {
-            write!(f, " ")?;
-
-            for &b in self.description() {
-                write!(f, "{}", b as char)?;
-            }
-        }
-
-        writeln!(f)?;
-
-        for &b in self.sequence() {
-            write!(f, "{}", b as char)?;
-        }
-
-        writeln!(f)?;
-
-        writeln!(f, "+")?;
-
-        for &b in self.quality_scores() {
-            write!(f, "{}", b as char)?;
-        }
-
-        writeln!(f)?;
-
-        Ok(())
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_fmt() {
-        let mut record = Record::new(Definition::new("r0", ""), "ATCG", "NDLS");
-        assert_eq!(record.to_string(), "@r0\nATCG\n+\nNDLS\n");
-
-        record.description_mut().extend_from_slice(b"LN:4");
-        assert_eq!(record.to_string(), "@r0 LN:4\nATCG\n+\nNDLS\n");
-    }
 
     #[test]
     fn test_clear() {